requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.14
django-ninja==0.24.0
django-channels==4.0.0
channels-redis==4.0.0
celery==5.5.0
redis==5.0.0
python-dotenv==1.0.0
pillow==11.0.0
django-crontab==0.7.1
pydicom==2.4.2
reportlab==3.6.12
openpyxl==3.1.2
